Update UI of login and registration pages
This commit includes the redesigning of the login and registration pages for enhanced user experience. Changes include, updating the language from English to German on the registration page, renaming and centralizing CSS styles for both pages, and improving the layout of form fields. The alert for registration errors is also aligned to the center now.